+++ src/mlpack/core/cv/metrics/r2_score_impl.hpp
@@ -33,24 +33,22 @@ double R2Score::Evaluate(MLAlgorithm& model,
   // Taking Predicted Output from the model.
   model.Predict(data, predictedResponses);
   // Mean value of response.
-  double mean_responses = arma::mean(responses);
+  double meanResponses = arma::mean(responses);
 
   // Calculate the numerator i.e. residual sum of squares.
-  double ss_res = arma::accu(arma::square(responses - predictedResponses));
+  double residualSumSquared = arma::accu(arma::square(responses -
+      predictedResponses));
 
   // Calculate the denominator i.e.total sum of squares.
-  double ss_tot = arma::accu(arma::square(responses - mean_responses));
+  double totalSumSquared = arma::accu(arma::square(responses - meanResponses));
 
-  // Handling undefined R2Score when both denominator and numerator is 0.0
-  if (ss_res == 0.0)
+  // Handling undefined R2 Score when both denominator and numerator is 0.0.
+  if (residualSumSquared == 0.0)
   {
-    if (ss_tot != 0.0)
-      return 1.0;
-    else
-      return DBL_MIN;
+    return totalSumSquared ? 1.0 : DBL_MIN;
   }
 
-  return 1 - ss_res / ss_tot;
+  return 1 - residualSumSquared / totalSumSquared;
 }
 
 } // namespace cv